Revert "slot: Add customizable `mergeProps` (#3953)"
This reverts commit 8d72d5f64914db82c12358a9ba9f4845b4985726.
