+++ apps/ssr-testing/app/rsc/page.tsx
@@ -7,14 +7,6 @@ import { AspectRatio, Label, Separator, Slot, VisuallyHidden } from 'radix-ui';
 const ServerSlot = Slot.createSlot('SsrTest.Slot');
 const ServerSlottable = Slot.createSlottable('SsrTest.Slottable');
 
-// A consumer-defined `mergeProps` passed directly to a Slot must still work in a
-// Server Component. The per-instance prop bypasses context (which is client
-// only), so it applies during server render just like on the client.
-const serverMergeProps: Slot.MergePropsFunction = (slotProps, childProps) => ({
-  ...Slot.mergeProps(slotProps, childProps),
-  'data-server-merged': 'true',
-});
-
 export default function Page() {
   return (
     <div>
@@ -35,15 +27,6 @@ export default function Page() {
         <span>aspect ratio content</span>
       </AspectRatio.Root>
       <VisuallyHidden.Root>hidden text</VisuallyHidden.Root>
-
-      {/*
-        Rendering a Slot in a Server Component: `useComposedRefs` relies only on
-        `useCallback` (available in the server build), and the per-instance
-        `mergeProps` prop is honored without a client boundary.
-      */}
-      <ServerSlot mergeProps={serverMergeProps} className="from-slot">
-        <a href="/">consumer-defined mergeProps applied on the server</a>
-      </ServerSlot>
     </div>
   );
 }
