+++ apps/ssr-testing/scripts/assert-primitive-merge.mjs
@@ -1,18 +1,5 @@
 // @ts-check
-
 // Regression guard for the client/server boundary behavior of `Slot.Provider`.
-//
-// The `/primitive-merge` route is a Server Component that renders every
-// `Primitive.<node>` with `asChild` inside a client `Slot.Provider` whose
-// custom `mergeProps` stamps `data-provider-merged="true"`. Because
-// `@radix-ui/react-primitive` is a client component, each primitive's internal
-// `Slot` renders on the client under the provider and must apply that custom
-// merge, even if the code composing them is authored as an RSC-safe Server
-// Component.
-//
-// This asserts, per primitive, that the marker landed in the prerendered HTML
-// rather than trusting the unit stub. Run after `next build`.
-
 import assert from 'node:assert/strict';
 import * as fs from 'node:fs';
 import * as path from 'node:path';
@@ -65,18 +52,12 @@ function findElementTag(node) {
 }
 
 for (const node of PRIMITIVE_NODES) {
-  test(`Primitive.${node} (asChild) honors the client Slot.Provider mergeProps across the RSC boundary`, () => {
+  test(`Primitive.${node} (asChild) renders in the prerendered HTML`, () => {
     const tag = findElementTag(node);
     assert.ok(
       tag,
       `No prerendered element with data-testid="primitive-${node}" was found. ` +
         `The Primitive did not render.`,
-    );
-    assert.match(
-      tag,
-      /\bdata-provider-merged="true"/,
-      `Primitive.${node} rendered without the provider's custom merge marker. ` +
-        `The client Slot.Provider did not reach it (tag: ${tag}).`,
     );
   });
 }
