+++ apps/storybook/stories/label.stories.tsx
@@ -1,6 +1,5 @@
-import { Label as LabelPrimitive, Slot } from 'radix-ui';
+import { Label as LabelPrimitive } from 'radix-ui';
 import styles from './label.stories.module.css';
-import { customMergeProps } from './custom-merge-props';
 
 const Label = LabelPrimitive.Root;
 
@@ -39,11 +38,3 @@ const Control = (props: any) => {
     </button>
   );
 };
-
-export const WithCustomMergeProps = () => (
-  <Slot.Provider mergeProps={customMergeProps}>
-    <Label className={styles.root} asChild>
-      <span data-custom-merge>Label (asChild)</span>
-    </Label>
-  </Slot.Provider>
-);
