+++ apps/storybook/stories/menubar.stories.tsx
@@ -1,9 +1,8 @@
 import * as React from 'react';
-import { Dialog, Menubar, Slot } from 'radix-ui';
+import { Dialog, Menubar } from 'radix-ui';
 import { foodGroups } from '@repo/test-data/foods';
 import styles from './menubar.stories.module.css';
 import { ExternalOverlayTrigger } from './external-overlay';
-import { customMergeProps } from './custom-merge-props';
 
 const subTriggerClass = [styles.item, styles.subTrigger].join(' ');
 
@@ -275,25 +274,6 @@ export const DismissesOnlyMenubarInsideDialog = () => {
     </div>
   );
 };
-export const WithCustomMergeProps = () => (
-  <Slot.Provider mergeProps={customMergeProps}>
-    <Menubar.Root className={styles.root}>
-      <Menubar.Menu>
-        <Menubar.Trigger className={styles.trigger} asChild>
-          <button data-custom-merge>File (asChild)</button>
-        </Menubar.Trigger>
-        <Menubar.Portal>
-          <Menubar.Content className={styles.content} sideOffset={5}>
-            <Menubar.Item className={styles.item} asChild>
-              <div data-custom-merge>New Tab (asChild)</div>
-            </Menubar.Item>
-            <Menubar.Item className={styles.item}>New Window</Menubar.Item>
-          </Menubar.Content>
-        </Menubar.Portal>
-      </Menubar.Menu>
-    </Menubar.Root>
-  </Slot.Provider>
-);
 
 export const Cypress = () => {
   const [loop, setLoop] = React.useState(false);
