+++ packages/react/slot/package.json
@@ -35,7 +35,6 @@
     "build": "radix-build"
   },
   "dependencies": {
-    "@radix-ui/primitive": "workspace:*",
     "@radix-ui/react-compose-refs": "workspace:*"
   },
   "devDependencies": {
