+++ packages/react/slot/src/index.ts
@@ -1,14 +1,9 @@
 export {
   Slot,
   Slottable,
-  SlotProvider,
   //
   Root,
-  Provider,
-  //
   createSlot,
   createSlottable,
-  //
-  mergeProps,
 } from './slot';
-export type { SlotProps, MergePropsFunction } from './slot';
+export type { SlotProps } from './slot';
